% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{calculate_null_dist}
\alias{calculate_null_dist}
\title{Calculated the simulated null distribution}
\usage{
calculate_null_dist(z_array, threshold = 1.5, binned = TRUE)
}
\arguments{
\item{z_array}{A simulation-by-time-by-predictor 3D array of z-values}

\item{threshold}{z-value threshold. Defaults to \code{1.5}.}

\item{binned}{Whether time points have been aggregated. Defaults to \code{TRUE}}
}
\description{
Calculated the simulated null distribution
}
\seealso{
clusterperm
}
