% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{cluster_permutation_tidiers}
\alias{cluster_permutation_tidiers}
\alias{tidy.timewise_statistics}
\alias{tidy.empirical_clusters}
\alias{tidy.null_cluster_dists}
\title{Tidiers for cluster permutation test objects}
\usage{
\method{tidy}{timewise_statistics}(x, ...)

\method{tidy}{empirical_clusters}(x, ...)

\method{tidy}{null_cluster_dists}(x, ...)
}
\arguments{
\item{x}{A \code{timewise_statistics}, \code{empirical_clusters}, or \code{null_clusters} object}

\item{...}{Unused}
}
\description{
Tidiers for cluster permutation test objects
}
