% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusters.R
\name{clusters_pvalue}
\alias{clusters_pvalue}
\title{Calculate bootstrapped p-values of clusters}
\usage{
clusters_pvalue(empirical_clusters, null_clusters)
}
\arguments{
\item{empirical_clusters}{A \code{empirical_clusters} object}

\item{null_clusters}{A \code{null_clusters} object}
}
\description{
Calculate bootstrapped p-values of clusters
}
\seealso{
\code{\link[=compute_empirical_clusters]{compute_empirical_clusters()}}, \code{\link[=compute_largest_null_clusters]{compute_largest_null_clusters()}}
}
