% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusters.R
\name{compute_empirical_clusters}
\alias{compute_empirical_clusters}
\title{Detect largest clusters from the time series data}
\usage{
compute_empirical_clusters(
  t_matrix,
  threshold = 1.5,
  binned = TRUE,
  top_n = 1L
)
}
\arguments{
\item{t_matrix}{A predictor-by-time matrix of cluster statistics.}

\item{threshold}{z-value threshold. Defaults to \code{1.5}.}

\item{binned}{Whether time points have been aggregated. Defaults to \code{TRUE} which allows length-1 clusters.}

\item{top_n}{How many clusters to return, ordered by the size of the cluster statistic.
Defaults to \code{1L} which returns the largest cluster. Use \code{NULL} to return all clusters.}
}
\description{
Detect largest clusters from the time series data
}
\seealso{
\code{\link[=jlmer_by_time]{jlmer_by_time()}}
}
