% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_timewise_statistics.R
\name{compute_timewise_statistics}
\alias{compute_timewise_statistics}
\title{Fit Julia mixed models to each time point of a time series data}
\usage{
compute_timewise_statistics(
  jlmer_spec,
  family = c("gaussian", "binomial"),
  ...
)
}
\arguments{
\item{jlmer_spec}{Data prepped for jlmer from \code{make_jlmer_spec()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{...}{Optional arguments to fit. Defaults to \code{fast = TRUE} and \code{progress = FALSE}.}
}
\value{
A predictor-by-time matrix of cluster statistics.
}
\description{
Fit Julia mixed models to each time point of a time series data
}
\seealso{
\code{\link[=jlmer]{jlmer()}}, \code{\link[=make_jlmer_spec]{make_jlmer_spec()}}
}
