% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{get_permuted_data_at}
\alias{get_permuted_data_at}
\title{Get the full permuted data sampled from the null}
\usage{
get_permuted_data_at(jlmer_spec, null_clusters, at)
}
\arguments{
\item{jlmer_spec}{Data prepped for jlmer from \code{make_jlmer_spec()}}

\item{null_clusters}{A \code{null_clusters} object.}

\item{at}{A vector of \code{.counter} values from \code{tidy(null_clusters)}.}
}
\value{
A long dataframe of permuted data from \code{null_clusters} specified in \code{at}.
}
\description{
Get the full permuted data sampled from the null
}
