% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{jlmer}
\alias{jlmer}
\alias{jlmer.fit}
\title{Fit a mixed model in Julia}
\usage{
jlmer(formula, data, family = c("gaussian", "binomial"), ...)

jlmer.fit(julia_formula, model_matrix, family = c("gaussian", "binomial"), ...)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Dataframe}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{...}{Optional arguments to fit (\code{fast}, \code{nAGQ}, etc.)}

\item{julia_formula}{From output of \code{jlmer_model_matrix()}}

\item{model_matrix}{From output of \code{jlmer_model_matrix()}}
}
\description{
Fit a mixed model in Julia
}
