% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{julia_model_tidiers}
\alias{julia_model_tidiers}
\alias{tidy.jlmer_mod}
\alias{glance.jlmer_mod}
\title{Tidier methods for Julia regression models}
\usage{
\method{tidy}{jlmer_mod}(x, effects = c("var_model", "ran_pars", "fixed"), ...)

\method{glance}{jlmer_mod}(x, ...)
}
\arguments{
\item{x}{An object of class \code{jlmer_mod}}

\item{effects}{One of "var_model", "ran_pars", or "fixed"}

\item{...}{Unused}
}
\description{
Tidier methods for Julia regression models
}
