% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-utils-exported.R
\name{julia_progress}
\alias{julia_progress}
\title{Set/get options for Julia progress bar}
\usage{
julia_progress(show, width)
}
\arguments{
\item{show}{Whether to show the progress bar. You may also pass in a list of \code{"show"} and \code{"width"}.}

\item{width}{Width of the progress bar.}
}
\value{
Previous values for \code{show} and \code{width}
}
\description{
Set/get options for Julia progress bar
}
\examples{
\dontrun{
jlmerclusterperm_setup()

# Set Julia progress options and save old state
old_progress_opts <- julia_progress(show = FALSE, width = 30)
old_progress_opts

# Restored old state
julia_progress(old_progress_opts)
identical(old_progress_opts, julia_progress())
}

}
