% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{julia_setup_ok}
\alias{julia_setup_ok}
\title{Check Julia setup requirements for jlmerclusterperm}
\usage{
julia_setup_ok()
}
\value{
Boolean
}
\description{
Check Julia setup requirements for jlmerclusterperm
}
\examples{
julia_setup_ok()
}
