% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{largest_clusters}
\alias{largest_clusters}
\title{Calculate largest clusters by predictor from a z-matrix}
\usage{
largest_clusters(z_matrix, threshold = 1.5, binned = TRUE)
}
\arguments{
\item{z_matrix}{A predictor-by-time matrix of z-values}

\item{threshold}{z-value threshold. Defaults to \code{1.5}.}

\item{binned}{Whether time points have been aggregated. Defaults to \code{TRUE}}
}
\description{
Calculate largest clusters by predictor from a z-matrix
}
\seealso{
jlmer_by_time
}
