% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer_spec.R
\name{make_jlmer_spec}
\alias{make_jlmer_spec}
\title{Construct a model matrix and expand model formula suitable for MixedModels and lme4}
\usage{
make_jlmer_spec(
  formula,
  data,
  subject = NULL,
  trial = NULL,
  time = NULL,
  drop_terms = NULL
)
}
\arguments{
\item{formula}{Model formula in R lme4 syntax}

\item{data}{Dataframe}

\item{subject}{Column for subjects or participants in the data.}

\item{trial}{Column for trials in the data. Must uniquely identify a time series within subject
(for example, the items in a counterbalanced design where each subject sees exactly one item).}

\item{time}{Column for time in the data.}

\item{drop_terms}{Terms to drop from the new model formula}
}
\value{
An object of class \code{jlmer_spec}
}
\description{
Construct a model matrix and expand model formula suitable for MixedModels and lme4
}
