% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{prep_jlmer_data}
\alias{prep_jlmer_data}
\title{Construct a model matrix and expand model formula suitable for MixedModels and lme4}
\usage{
prep_jlmer_data(
  fm,
  df,
  subject = NULL,
  item = NULL,
  time = NULL,
  drop_terms = NULL
)
}
\arguments{
\item{fm}{Formula}

\item{df}{Dataframe}

\item{subject}{Column for subjects in the data}

\item{item}{Column for items in the data}

\item{time}{Column for time in the data}

\item{drop_terms}{Terms to drop from the new model formula}
}
\value{
A list of R formula for lme4, Julia formula for MixedModels, and the model matrix as a data frame
}
\description{
Construct a model matrix and expand model formula suitable for MixedModels and lme4
}
