% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop-utils.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{stopJulia}
\alias{tidy}
\alias{glance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{JuliaConnectoR}{\code{\link[JuliaConnectoR]{stopJulia}}}
}}

