% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_counter.R
\name{rng_counter}
\alias{rng_counter}
\alias{set_rng_counter}
\alias{reset_rng_counter}
\title{RNG seed and counter for the permutation algorithm}
\usage{
set_rng_counter(i)

reset_rng_counter()
}
\arguments{
\item{i}{Counter number}
}
\description{
RNG seed and counter for the permutation algorithm
}
\keyword{internal}
