% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvalue.R
\name{calculate_clusters_pvalues}
\alias{calculate_clusters_pvalues}
\alias{clusters_are_comparable}
\title{Calculate bootstrapped p-values of clusters}
\usage{
calculate_clusters_pvalues(empirical_clusters, null_clusters, add1 = FALSE)

clusters_are_comparable(empirical_clusters, null_clusters)
}
\arguments{
\item{empirical_clusters}{A \code{empirical_clusters} object}

\item{null_clusters}{A \code{null_clusters} object}

\item{add1}{Whether to add 1 to the numerator before calculating probability.
Defaults to \code{FALSE}. Use \code{TRUE} to effectively include the observed statistic
as part of the null distribution (recommended with larger \code{nsim} prior to publishing results).}
}
\description{
Calculate bootstrapped p-values of clusters
}
\seealso{
\code{\link[=extract_empirical_clusters]{extract_empirical_clusters()}}, \code{\link[=extract_null_cluster_dists]{extract_null_cluster_dists()}}
}
