% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvalue.R
\name{calculate_clusters_pvalues}
\alias{calculate_clusters_pvalues}
\alias{clusters_are_comparable}
\title{Calculate bootstrapped p-values of cluster-mass statistics}
\usage{
calculate_clusters_pvalues(
  empirical_clusters,
  null_cluster_dists,
  add1 = FALSE
)

clusters_are_comparable(empirical_clusters, null_cluster_dists, error = FALSE)
}
\arguments{
\item{empirical_clusters}{A \code{empirical_clusters} object}

\item{null_cluster_dists}{A \code{null_cluster_dists} object}

\item{add1}{Whether to add 1 to the numerator and denominator when calculating the p-value.
Use \code{TRUE} to effectively count the observed statistic as part of the permuted
null distribution (recommended with larger \code{nsim} prior to publishing results).}

\item{error}{Whether to throw an error if incompatible}
}
\value{
An \code{empirical_clusters} object augmented with p-values.
}
\description{
Calculate bootstrapped p-values of cluster-mass statistics
}
\examples{
library(dplyr, warn.conflicts = FALSE)
jlmerclusterperm_setup(restart = FALSE, verbose = FALSE)

# Specification object
spec <- make_jlmer_spec(
  weight ~ 1 + Diet, filter(ChickWeight, Time <= 20),
  subject = "Chick", time = "Time"
)
spec

# Make empirical clusters
empirical_statistics <- compute_timewise_statistics(spec)
empirical_clusters <- extract_empirical_clusters(empirical_statistics, threshold = 2)
empirical_clusters

# Make null cluster-mass distribution
reset_rng_state()
null_statistics <- permute_timewise_statistics(spec, nsim = 100)
null_cluster_dists <- extract_null_cluster_dists(null_statistics, threshold = 2)

# Significance test the empirical cluster(s) from each predictor against the simulated null
calculate_clusters_pvalues(empirical_clusters, null_cluster_dists)

# Set `add1 = TRUE` to normalize by adding 1 to numerator and denominator
calculate_clusters_pvalues(empirical_clusters, null_cluster_dists, add1 = TRUE)

# This sequence of procedures is equivalent to `clusterpermute()`
reset_rng_state()
clusterpermute(spec, threshold = 2, nsim = 100, progress = FALSE)

# The empirical clusters and the null cluster-mass distribution must be comparable
empirical_clusters2 <- extract_empirical_clusters(empirical_statistics, threshold = 3)
# Below code (not run) errors because thresholds are different (2 vs. 3)
# calculate_clusters_pvalues(empirical_clusters2, null_cluster_dists)

# Check for compatibility with `clusters_are_comparable()`
clusters_are_comparable(empirical_clusters, null_cluster_dists)
clusters_are_comparable(empirical_clusters2, null_cluster_dists)

}
\seealso{
\code{\link[=extract_empirical_clusters]{extract_empirical_clusters()}}, \code{\link[=extract_null_cluster_dists]{extract_null_cluster_dists()}}
}
