% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterpermute.R
\name{clusterpermute}
\alias{clusterpermute}
\title{Fit Julia mixed models to each time point of a time series data}
\usage{
clusterpermute(
  jlmer_data,
  family = c("gaussian", "binomial"),
  nsim = 100L,
  predictors = NULL,
  ...
)
}
\arguments{
\item{jlmer_data}{Data prepped for jlmer from \code{prep_for_jlmer()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{nsim}{Number of simulations description}

\item{predictors}{A subset of predictors to test. Defaults to \code{NULL} which tests all predictors.}

\item{...}{Optional arguments to fit. Defaults to \code{fast = TRUE} and \code{progress = FALSE}.}
}
\value{
A list of Predictor x Time matrix of z-values
}
\description{
Fit Julia mixed models to each time point of a time series data
}
\seealso{
prep_jlmer_data
}
