% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterpermute.R
\name{clusterpermute}
\alias{clusterpermute}
\title{Fit Julia mixed models to each time point of a time series data}
\usage{
clusterpermute(
  julia_formula,
  data,
  time,
  family = c("gaussian", "binomial"),
  nsim = 100L,
  threshold = 1.5,
  participant_col,
  binned = TRUE,
  ...
)
}
\arguments{
\item{julia_formula}{A formula in MixedModels.jl syntax}

\item{data}{DataFrame}

\item{time}{Column representing time in \code{data}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{nsim}{Number of simulations description}

\item{...}{Optional arguments to fit. Defaults to \code{fast = TRUE} and \code{progress = FALSE}.}
}
\value{
A list of Predictor x Time matrix of z-values
}
\description{
Fit Julia mixed models to each time point of a time series data
}
\seealso{
jlmer_model_matrix
}
