% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterpermute.R
\name{clusterpermute}
\alias{clusterpermute}
\title{Run a complete cluster-based permutation analysis}
\usage{
clusterpermute(
  jlmer_spec,
  family,
  statistic,
  threshold,
  nsim,
  predictors,
  binned,
  top_n,
  add1,
  ...
)
}
\arguments{
\item{jlmer_spec}{Data prepped for jlmer from \code{make_jlmer_spec()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{statistic}{Test statistic for calculating cluster mass.
Can be one of \code{"t"} (default) from the regression model output or
\code{"chisq"} from a likelihood ratio test (takes about twice as long to calculate).}

\item{threshold}{The threshold value that a statistic must pass to contribute to cluster mass.
Interpretation differs on the choice of statistic (more below):
\itemize{
\item If \code{statistic = "t"}, the threshold for t-value (beta/std.err) from the regression model.
\item If \code{statistic = "chisq"}, the threshold for the p-value of chi-squared statistics from likelihood ratio tests.
}}

\item{nsim}{Number of simulations description}

\item{predictors}{(Optional) a subset of predictors to test. Defaults to \code{NULL} which tests all predictors.}

\item{binned}{Whether the data has been aggregated/collapsed into time bins. Defaults to \code{FALSE},
which requires a cluster to span at least two time points. If \code{TRUE}, allows length-1 clusters to exist.}

\item{top_n}{How many clusters to return, in order of the size of the cluster statistic.
Defaults to \code{1} which only returns the largest cluster. Use \code{Inf} to return all possible clusters.}

\item{add1}{Whether to add 1 to the numerator before calculating probability.
Defaults to \code{FALSE}. Use \code{TRUE} to effectively include the observed statistic
as part of the null distribution (recommended with larger \code{nsim} prior to publishing results).}

\item{...}{Optional arguments passed to Julia for model fitting.
Defaults to \code{fast = TRUE} and \code{progress = FALSE}.}
}
\value{
An \code{empirical_clusters} object with p-values
}
\description{
Run a complete cluster-based permutation analysis
}
\seealso{
\code{\link[=compute_timewise_statistics]{compute_timewise_statistics()}}, \code{\link[=permute_timewise_statistics]{permute_timewise_statistics()}},
\code{\link[=extract_empirical_clusters]{extract_empirical_clusters()}}, \code{\link[=extract_null_cluster_dists]{extract_null_cluster_dists()}},
\code{\link[=calculate_clusters_pvalues]{calculate_clusters_pvalues()}}
}
