% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_timewise_statistics.R
\name{compute_timewise_statistics}
\alias{compute_timewise_statistics}
\title{Fit Julia regression models to each time point of a time series data}
\usage{
compute_timewise_statistics(
  jlmer_spec,
  family = c("gaussian", "binomial"),
  statistic = c("t", "chisq"),
  ...
)
}
\arguments{
\item{jlmer_spec}{Data prepped for jlmer from \code{make_jlmer_spec()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{statistic}{Test statistic for calculating cluster mass.
Can be one of \code{"t"} (default) from the regression model output or
\code{"chisq"} from a likelihood ratio test (takes about twice as long to calculate).}

\item{...}{Optional arguments passed to Julia for model fitting.
Defaults to \code{fast = TRUE} (when \code{family = "binomial"}) and \code{progress = FALSE}.}
}
\value{
A predictor-by-time matrix of cluster statistics, of class \code{timewise_statistics}.
}
\description{
Fit Julia regression models to each time point of a time series data
}
\seealso{
\code{\link[=jlmer]{jlmer()}}, \code{\link[=make_jlmer_spec]{make_jlmer_spec()}}
}
