% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clusters.R
\name{extract_null_cluster_dists}
\alias{extract_null_cluster_dists}
\title{Construct a distribution of the largest cluster statistics from bootstrapped permutations}
\usage{
extract_null_cluster_dists(t_array, threshold = 1.5, binned = TRUE)
}
\arguments{
\item{t_array}{A simulation-by-time-by-predictor 3D array of cluster statistics.}

\item{threshold}{Cluster statistic threshold. Defaults to \code{1.5}.}

\item{binned}{Whether time points have been aggregated. Defaults to \code{TRUE} which allows length-1 clusters.}
}
\description{
Construct a distribution of the largest cluster statistics from bootstrapped permutations
}
\seealso{
\code{\link[=clusterpermute]{clusterpermute()}}
}
