% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{jlmer}
\alias{jlmer}
\title{Fit a mixed model in Julia}
\usage{
jlmer(julia_formula, data)
}
\arguments{
\item{julia_formula}{Model formula in Julia syntax}

\item{data}{Dataframe}

\item{...}{Ignored.}
}
\description{
Fit a mixed model in Julia
}
