% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{jlmer}
\alias{jlmer}
\title{Fit a Julia mixed model}
\usage{
jlmer(jlmer_data, family = c("gaussian", "binomial"), ...)
}
\arguments{
\item{jlmer_data}{Data prepped for jlmer from \code{prep_for_jlmer()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{...}{Optional arguments to fit (\code{fast}, \code{nAGQ}, etc.)}
}
\description{
Fit a Julia mixed model
}
\seealso{
prep_jlmer_data
}
