% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{jlmer}
\alias{jlmer}
\title{Fit a Julia mixed model}
\usage{
jlmer(jlmer_spec, family = c("gaussian", "binomial"), ..., progress = FALSE)
}
\arguments{
\item{jlmer_spec}{Data prepped for jlmer from \code{make_jlmer_spec()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{...}{Optional arguments passed to Julia for model fitting.}

\item{progress}{If \code{TRUE}, prints the timing of iterations.}
}
\value{
A \code{jlmer_mod} object.
}
\description{
Fit a Julia mixed model
}
\seealso{
\code{\link[=make_jlmer_spec]{make_jlmer_spec()}}
}
