% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{jlmer_by_time}
\alias{jlmer_by_time}
\title{Fit Julia mixed models to each time point of a timeseries data}
\usage{
jlmer_by_time(
  julia_formula,
  data,
  time,
  family = c("gaussian", "binomial"),
  ...
)
}
\arguments{
\item{julia_formula}{A formula in MixedModels.jl syntax}

\item{data}{DataFrame}

\item{time}{The column representing time in the timeseries}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{...}{Optional arguments to fit. Defaults to \code{fast = TRUE} and \code{progress = FALSE}}
}
\value{
A Predictor x Time matrix of t-values
}
\description{
Fit Julia mixed models to each time point of a timeseries data
}
\seealso{
jlmer_model_matrix
}
