% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{jlmer_by_time}
\alias{jlmer_by_time}
\title{Fit Julia mixed models to each time point of a time series data}
\usage{
jlmer_by_time(jlmer_data, family = c("gaussian", "binomial"), ...)
}
\arguments{
\item{jlmer_data}{Data prepped for jlmer from \code{prep_for_jlmer()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{...}{Optional arguments to fit. Defaults to \code{fast = TRUE} and \code{progress = FALSE}.}
}
\value{
A Predictor x Time matrix of z-values
}
\description{
Fit Julia mixed models to each time point of a time series data
}
\seealso{
prep_jlmer_data
}
