% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmerclusterperm-package.R
\docType{package}
\name{jlmerclusterperm-package}
\alias{jlmerclusterperm}
\alias{jlmerclusterperm-package}
\title{jlmerclusterperm: Bootstrapped cluster-based permutation analysis for time series data with GLM.jl and MixedModels.jl}
\description{
A Julia-based implementation of fast cluster-based permutation tests for time series data using mixed-effects models. Provides a modular interface for the steps of calculating timewise statistics, extracting clusters, and calculating p-values. Calls MixedModels.jl and GLM.jl to fit regression models using JuliaConnectoR (requires Julia installation).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yjunechoe/jlmerclusterperm}
  \item \url{https://yjunechoe.github.io/jlmerclusterperm/}
  \item Report bugs at \url{https://github.com/yjunechoe/jlmerclusterperm/issues}
}

}
\author{
\strong{Maintainer}: June Choe \email{jchoe001@gmail.com} (\href{https://orcid.org/0000-0002-0701-921X}{ORCID})

}
\keyword{internal}
