% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmerclusterperm-package.R
\docType{package}
\name{jlmerclusterperm-package}
\alias{jlmerclusterperm}
\alias{jlmerclusterperm-package}
\title{jlmerclusterperm: Cluster-based Permutation Analysis for Densely Sampled Time Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of fast cluster-based permutation analysis (CPA) for densely-sampled time data developed in Maris & Oostenveld, 2007 \doi{10.1016/j.jneumeth.2007.03.024}. Supports (generalized, mixed-effects) regression models for the calculation of timewise statistics. Provides both a wholesale and a piecemeal interface to the CPA procedure with an emphasis on interpretability and diagnostics. Integrates 'Julia' libraries 'MixedModels.JL' and 'GLM.JL' for performance improvements, with additional functionalities for interfacing with 'Julia' from R powered by the 'JuliaConnectoR' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yjunechoe/jlmerclusterperm}
  \item \url{https://yjunechoe.github.io/jlmerclusterperm/}
  \item Report bugs at \url{https://github.com/yjunechoe/jlmerclusterperm/issues}
}

}
\author{
\strong{Maintainer}: June Choe \email{jchoe001@gmail.com} (\href{https://orcid.org/0000-0002-0701-921X}{ORCID}) [copyright holder]

}
\keyword{internal}
