% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{jlmerclusterperm_setup}
\alias{jlmerclusterperm_setup}
\title{Initial setup for the jlmerclusterperm package}
\usage{
jlmerclusterperm_setup(..., cache_dir = NULL, restart = TRUE, verbose = TRUE)
}
\arguments{
\item{...}{Ignored.}

\item{cache_dir}{The location to write out package cache files (namely, Manifest.toml).
If \code{NULL} (default), attempts to write to the package's cache directory discovered via
\code{R_user_dir()} and falls back to \code{tempdir()}.}

\item{restart}{Whether to set up a fresh Julia session, given that one is already running.
If \code{FALSE}, setup is skipped when the current Julia session is already set up for
jlmerclusterperm, and runs otherwise (e.g., if the session was stopped). Use
\code{jlmerclusterperm_setup(restart = FALSE)} at the top of a script to set up Julia only
if it is not already set up.}

\item{verbose}{Whether to print progress and messages from Julia in the console}
}
\value{
Invisibly, \code{TRUE} if Julia is set up for jlmerclusterperm.
}
\description{
Initial setup for the jlmerclusterperm package
}
\examples{
\dontshow{if (julia_setup_ok()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
options("jlmerclusterperm.nthreads" = 2)
jlmerclusterperm_setup(cache_dir = tempdir(), verbose = FALSE)

\dontshow{
JuliaConnectoR::stopJulia()
}
}
\dontshow{\}) # examplesIf}
}
