% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{jlmerclusterperm_setup}
\alias{jlmerclusterperm_setup}
\title{Initial setup for jlmerclusterperm package.}
\usage{
jlmerclusterperm_setup(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Initial setup for jlmerclusterperm package.
}
\examples{
\dontrun{
jlmerclusterperm_setup()
}

}
