% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{jlmerclusterperm_setup}
\alias{jlmerclusterperm_setup}
\title{Initial setup for jlmerclusterperm package.}
\usage{
jlmerclusterperm_setup(..., verbose = TRUE)
}
\arguments{
\item{...}{Ignored.}

\item{verbose}{Print progress and messages from Julia in the console}
}
\description{
Initial setup for jlmerclusterperm package.
}
\examples{
\dontrun{
jlmerclusterperm_setup()
}

}
