% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{jlmerclusterperm_setup}
\alias{jlmerclusterperm_setup}
\title{Initial setup for the jlmerclusterperm package}
\usage{
jlmerclusterperm_setup(..., restart = TRUE, verbose = TRUE)
}
\arguments{
\item{...}{Ignored.}

\item{restart}{Whether to set up a fresh Julia session, given that one is already running.
If \code{FALSE} and \code{jlmerclusterperm_setup()} has already been called, nothing happens.}

\item{verbose}{Print progress and messages from Julia in the console}
}
\value{
TRUE
}
\description{
Initial setup for the jlmerclusterperm package
}
\examples{
\dontshow{if (JuliaConnectoR::juliaSetupOk()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
options("jlmerclusterperm.nthreads" = 2)
jlmerclusterperm_setup(verbose = FALSE)

\dontshow{
JuliaConnectoR::stopJulia()
}
}
\dontshow{\}) # examplesIf}
}
