% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_rng.R
\name{julia_rng}
\alias{julia_rng}
\alias{set_rng_state}
\alias{reset_rng_state}
\alias{get_rng_state}
\title{Interface to the RNG counter}
\usage{
set_rng_state(i)

reset_rng_state()

get_rng_state()
}
\arguments{
\item{i}{Counter number}
}
\description{
Interface to the RNG counter
}
\keyword{internal}
