% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_counter.R
\name{julia_rng}
\alias{julia_rng}
\alias{set_rng_counter}
\alias{reset_rng_counter}
\alias{set_rng_seed}
\title{RNG seed and counter for the permutation algorithm}
\usage{
set_rng_counter(counter)

reset_rng_counter()

set_rng_seed(seed)
}
\arguments{
\item{i}{Counter number}
}
\description{
RNG seed and counter for the permutation algorithm
}
\keyword{internal}
