% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer_spec.R
\name{make_jlmer_spec}
\alias{make_jlmer_spec}
\title{Create a specifications object for fitting regression models in Julia}
\usage{
make_jlmer_spec(
  formula,
  data,
  subject = NULL,
  trial = NULL,
  time = NULL,
  drop_terms = NULL
)
}
\arguments{
\item{formula}{Model formula in R syntax}

\item{data}{A data frame}

\item{subject}{Column for subjects in the data.}

\item{trial}{Column for trials in the data. Must uniquely identify a time series within subject
(for example, the column for items in a counterbalanced design where each subject sees exactly one item).}

\item{time}{Column for time in the data.}

\item{drop_terms}{(Optional) any terms to drop from the reconstructed model formula}
}
\value{
An object of class \code{jlmer_spec}.
}
\description{
Create a specifications object for fitting regression models in Julia
}
