% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute_by_predictor}
\alias{permute_by_predictor}
\title{Permute data by predictor}
\usage{
permute_by_predictor(
  jlmer_data,
  predictors,
  predictor_type = c("guess", "between_participant", "within_participant"),
  n = 1L
)
}
\arguments{
\item{jlmer_data}{Data prepped for jlmer from \code{prep_for_jlmer()}}

\item{predictors}{A vector of terms from the model. If multiple, the must form the levels of one predictor.}

\item{predictor_type}{Whether the predictor is between or within participant. Default is to guess.}

\item{n}{Number of permuted data to generate}
}
\value{
A long dataframe with \code{.id} column representing replication IDs.
}
\description{
Permute data by predictor
}
