% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_timewise_statistics.R
\name{permute_timewise_statistics}
\alias{permute_timewise_statistics}
\title{Fit Julia mixed models to each time point of a time series data}
\usage{
permute_timewise_statistics(
  jlmer_spec,
  family = c("gaussian", "binomial"),
  nsim = 100L,
  predictors = NULL,
  ...
)
}
\arguments{
\item{jlmer_spec}{Data prepped for jlmer from \code{make_jlmer_spec()}}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{nsim}{Number of simulations description}

\item{predictors}{A subset of predictors to test. Defaults to \code{NULL} which tests all predictors.}

\item{...}{Optional arguments to fit. Defaults to \code{fast = TRUE} and \code{progress = FALSE}.}
}
\value{
A simulation-by-time-by-predictor 3D array of cluster statistics.
}
\description{
Fit Julia mixed models to each time point of a time series data
}
\seealso{
\code{\link[=make_jlmer_spec]{make_jlmer_spec()}}
}
