% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{to_jlmer}
\alias{to_jlmer}
\title{Fit a Julia mixed model using lme4 syntax}
\usage{
to_jlmer(
  formula,
  data,
  family = c("gaussian", "binomial"),
  jlmer_spec_opts = list(),
  ...
)
}
\arguments{
\item{formula}{Model formula in R syntax}

\item{data}{A data frame}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{jlmer_spec_opts}{Other options passed to \code{make_jlmer_spec()}}

\item{...}{Optional arguments passed to Julia for model fitting.}
}
\value{
A \code{jlmer_mod} object.
}
\description{
Fit a Julia mixed model using lme4 syntax
}
\seealso{
\code{\link[=jlmer]{jlmer()}}, \code{\link[=make_jlmer_spec]{make_jlmer_spec()}}
}
