% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlmer.R
\name{to_jlmer}
\alias{to_jlmer}
\title{Fit a Julia mixed model using lmer syntax}
\usage{
to_jlmer(
  formula,
  data,
  family = c("gaussian", "binomial"),
  reformulate_opts = list(),
  ...
)
}
\arguments{
\item{formula}{Model formula in lme4 syntax}

\item{data}{Dataframe}

\item{family}{A GLM family. Currently supports "gaussian" and "binomial".}

\item{reformulate_opts}{Other options passed to \code{jlmer_model_matrix()}}

\item{...}{Optional arguments to fit (\code{fast}, \code{nAGQ}, etc.)}
}
\description{
Fit a Julia mixed model using lmer syntax
}
\seealso{
jlmer_model_matrix
}
