% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_search.R
\name{walk_threshold_steps}
\alias{walk_threshold_steps}
\title{Test cluster-mass probabilities over a range of threshold values}
\usage{
walk_threshold_steps(
  empirical_statistics,
  null_statistics,
  threshold_steps,
  top_n = Inf,
  binned = FALSE,
  add1 = TRUE
)
}
\arguments{
\item{empirical_statistics}{A predictor-by-time matrix of empirical timewise statistics.}

\item{null_statistics}{A simulation-by-time-by-predictor 3D array of null (permuted) timewise statistics.}

\item{threshold_steps}{A vector of threshold values to test}

\item{top_n}{How many clusters to return, in order of the size of the cluster statistic.
Defaults to \code{1} which only returns the largest cluster. Use \code{Inf} to return all possible clusters.}

\item{binned}{Whether the data has been aggregated/collapsed into time bins. Defaults to \code{FALSE},
which requires a cluster to span at least two time points. If \code{TRUE}, allows length-1 clusters to exist.}

\item{add1}{Whether to add 1 to the numerator before calculating probability.
Use \code{TRUE} to effectively count the observed statistic as part of the permuted
null distribution (recommended with larger \code{nsim} prior to publishing results).}
}
\value{
A data frame of clusters-mass statistics
}
\description{
Test cluster-mass probabilities over a range of threshold values
}
